SBOT-1967: refactoring, package name edit
