--- src/main/java/ru/idfedorov09/telegram/bot/SnoBotHandlerApplication.java
@@ -1,19 +1,0 @@
-package ru.idfedorov09.telegram.bot;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.annotation.EnableScheduling;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-@SpringBootApplication
-@EnableTransactionManagement
-@EnableAsync
-@EnableScheduling
-public class SnoBotHandlerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SnoBotHandlerApplication.class, args);
-	}
-
-}
