+++ src/main/java/ru/idfedorov09/telegram/bot/base/UpdatesHandler.java
@@ -1,0 +1,10 @@
+package ru.idfedorov09.telegram.bot.base;
+
+import org.telegram.telegrambots.bots.TelegramLongPollingBot;
+import org.telegram.telegrambots.meta.api.objects.Update;
+
+
+public interface UpdatesHandler {
+
+    void handle(TelegramLongPollingBot telegramLongPollingBot, Update update);
+}
