+++ src/main/java/ru/idfedorov09/telegram/bot/base/data/GlobalConstants.kt
@@ -1,0 +1,11 @@
+package ru.idfedorov09.telegram.bot.base.data
+
+object GlobalConstants {
+
+    const val HEALTH_PATH = "/is_alive"
+    const val WEBHOOK_PATH = "/"
+
+    const val QUALIFIER_FLOW_TG_BOT = "Example flow"
+    const val QUALIFIER_FLOW_HEALTH_STATUS = "Example health-check flow"
+    const val QUEUE_PRE_PREFIX = "frjekcs_ewer_idfed09_user_bot_que_"
+}
