+++ src/main/java/ru/idfedorov09/telegram/bot/base/data/model/HealthStatus.kt
@@ -1,0 +1,10 @@
+package ru.idfedorov09.telegram.bot.base.data.model
+
+data class HealthStatus(
+    val generalStatus: String,
+) {
+    object Messages {
+        const val HEALTH_FLOW_NOT_CONTAINS_STATUS = "Error during flow execution: " +
+            "the context does not contain information about the status of the application."
+    }
+}
