+++ src/main/java/ru/idfedorov09/telegram/bot/base/executor/Executor.kt
@@ -1,0 +1,33 @@
+package ru.idfedorov09.telegram.bot.base.executor
+
+import jakarta.annotation.PostConstruct
+import org.slf4j.LoggerFactory
+import org.springframework.stereotype.Component
+import org.telegram.telegrambots.bots.TelegramLongPollingBot
+import org.telegram.telegrambots.meta.api.objects.Update
+import ru.idfedorov09.telegram.bot.base.config.BotContainer
+
+@Component
+class Executor(
+    private val botContainer: BotContainer,
+) : TelegramLongPollingBot() {
+
+    companion object {
+        private val log = LoggerFactory.getLogger(Executor::class.java)
+    }
+
+    @PostConstruct
+    fun postConstruct() {
+        log.info("Telegram method executor created.")
+    }
+
+    override fun onUpdateReceived(update: Update) {}
+
+    override fun getBotUsername(): String {
+        return botContainer.BOT_NAME
+    }
+
+    override fun getBotToken(): String {
+        return botContainer.BOT_TOKEN
+    }
+}
