+++ src/main/java/ru/idfedorov09/telegram/bot/base/fetchers/bot/TestFetcher.kt
@@ -1,0 +1,34 @@
+package ru.idfedorov09.telegram.bot.base.fetchers.bot
+
+import org.slf4j.LoggerFactory
+import org.springframework.stereotype.Component
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage
+import org.telegram.telegrambots.meta.api.objects.Update
+import ru.idfedorov09.telegram.bot.base.executor.Executor
+import ru.idfedorov09.telegram.bot.base.util.UpdatesUtil
+import ru.mephi.sno.libs.flow.belly.InjectData
+import ru.mephi.sno.libs.flow.fetcher.GeneralFetcher
+
+@Component
+class TestFetcher(
+    private val updatesUtil: UpdatesUtil,
+    private val bot: Executor,
+) : GeneralFetcher() {
+
+    companion object {
+        private val log = LoggerFactory.getLogger(TestFetcher::class.java)
+    }
+
+    @InjectData
+    fun doFetch(
+        update: Update,
+    ) {
+        log.info("UPDATE RECEIVE: {}", update)
+        bot.execute(
+            SendMessage(
+                updatesUtil.getChatId(update)!!,
+                "ok!!",
+            ),
+        )
+    }
+}
