+++ src/main/java/ru/idfedorov09/telegram/bot/base/flow/ExpContainer.kt
@@ -1,0 +1,11 @@
+package ru.idfedorov09.telegram.bot.base.flow
+
+import ru.mephi.sno.libs.flow.belly.Mutable
+
+/**
+ * Объект контекста флоу, содержащий информацию о работающих фичах, режимах и тд и тп
+ */
+@Mutable
+data class ExpContainer(
+    val test: String = "test"
+)
