+++ src/main/java/ru/idfedorov09/telegram/bot/base/flow/TelegramBotFlowConfiguration.kt
@@ -1,0 +1,32 @@
+package ru.idfedorov09.telegram.bot.base.flow
+
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import ru.idfedorov09.telegram.bot.base.data.GlobalConstants.QUALIFIER_FLOW_TG_BOT
+import ru.idfedorov09.telegram.bot.base.fetchers.bot.TestFetcher
+import ru.mephi.sno.libs.flow.belly.FlowBuilder
+
+/**
+ * Основной класс, в котором строится последовательность вычислений (граф) для бота
+ */
+@Configuration
+open class TelegramBotFlowConfiguration(
+    private val testFetcher: TestFetcher,
+) {
+
+    /**
+     * Возвращает построенный граф; выполняется только при запуске приложения
+     */
+    @Bean(QUALIFIER_FLOW_TG_BOT)
+    open fun flowBuilder(): FlowBuilder {
+        val flowBuilder = FlowBuilder()
+        flowBuilder.buildFlow()
+        return flowBuilder
+    }
+
+    private fun FlowBuilder.buildFlow() {
+        group {
+            fetch(testFetcher)
+        }
+    }
+}
