+++ src/main/java/ru/idfedorov09/telegram/bot/base/util/RedisUtil.kt
@@ -1,0 +1,14 @@
+package ru.idfedorov09.telegram.bot.base.util
+
+import redis.clients.jedis.Jedis
+import ru.idfedorov09.telegram.bot.base.data.model.RedisServerData
+
+object RedisUtil {
+    fun getConnection(redisServerData: RedisServerData) = redisServerData.run {
+        Jedis(host, port).also { jedis ->
+            runCatching {
+                password?.let { jedis.auth(it) }
+            }
+        }
+    }
+}
