--- src/main/java/ru/idfedorov09/telegram/bot/data/model/RedisServerData.kt
@@ -1,11 +1,0 @@
-package ru.idfedorov09.telegram.bot.data.model
-
-data class RedisServerData(
-    val port: Int,
-    val host: String,
-
-    /**
-     * hint: если пароль не указан, пробуем подключиться без него
-     */
-    val password: String?,
-)
